% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-SummarizedExperiment.R
\name{contrastResultsNames}
\alias{contrastResultsNames}
\title{Extract results from a DESeq analysis}
\usage{
contrastResultsNames(object)
}
\arguments{
\item{object}{A \linkS4class{SummarizedExperiment} object.}
}
\value{
For \code{contrastResultsNames}: the names of embedded contrast results available.
}
\description{
\code{contrastResultsNames} returns the names of contrast results embedded in \code{object}.
}
\examples{
library("iSEEde")
library("airway")
library("DESeq2")
library("iSEE")

##
# Example data ----
##

data("airway")
airway$dex <- relevel(airway$dex, "untrt")

dds <- DESeqDataSet(airway, ~ 0 + dex + cell)

dds <- DESeq(dds)
res_deseq2 <- results(dds, contrast = list("dextrt", "dexuntrt"))
airway <- embedContrastResults(res_deseq2, airway, name = "dex: trt vs untrt")

##
# Demo ---
##

contrastResultsNames(airway)
}
