% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{de-generics}
\alias{de-generics}
\alias{pvalue}
\alias{log2foldchange}
\alias{log2average}
\title{Generics for Differential Expression Results}
\description{
An overview of the generics for accessing common pieces of information in differential expression results.
}
\section{Definitions}{

\itemize{
\item \code{pvalue(x)} returns a numeric vector of raw p-values.
\item \code{log2foldchange(x)} returns a numeric vector of log2-fold-change values.
\item \code{log2average(x)} returns a numeric vector of average log2-expression values.
}
}

\author{
Kevin Rue-Albrecht
}
