% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETable-class.R
\docType{methods}
\name{DETable-class}
\alias{DETable-class}
\alias{DETable}
\alias{initialize,DETable-method}
\alias{.cacheCommonInfo,DETable-method}
\alias{.createObservers,DETable-method}
\alias{.defineDataInterface,DETable-method}
\alias{.fullName,DETable-method}
\alias{.generateTable,DETable-method}
\alias{.panelColor,DETable-method}
\alias{.refineParameters,DETable-method}
\title{The DETable class}
\description{
The DETable class is a \linkS4class{RowTable} subclass that is dedicated to creating a volcano plot.
It retrieves the table of results for the selected differential expression contrast and creates an interactive table where each row represents a feature.
}
\section{Slot overview}{

The following slots control the test procedure:
\itemize{
\item \code{ContrastName}, a character scalar indicating the name of the contrast to display.
\item \code{RoundDigits}, a logical scalar indicating whether to round numeric values (see \code{SignifDigits}).
\item \code{SignifDigits}, an integer scalar indicating the number of significant digits to use for rounding numbers  (see \code{RoundDigits}).
}

In addition, this class inherits all slots from its parent \linkS4class{RowTable} and \linkS4class{Table} classes.
}

\examples{
x <- DETable()
x
}
