% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEde-package.R
\docType{package}
\name{iSEEde-pkg}
\alias{iSEEde}
\alias{iSEEde-package}
\title{iSEE extension for panels related to differential expression analysis}
\description{
\code{iSEEde} is a package that provides panels for \pkg{iSEE},
facilitating the interactive visualisation of differential expression results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iSEE/iSEEde}
  \item Report bugs at \url{https://support.bioconductor.org/t/iSEEde}
}

}
\author{
Kevin Rue-Albrecht \email{kevin.rue-albrecht@imm.ox.ac.uk}
}
\keyword{internal}
