% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot-class.R
\docType{methods}
\name{MAPlot-class}
\alias{MAPlot-class}
\alias{MAPlot}
\title{The MAPlot class}
\description{
The MAPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating a volcano plot.
It retrieves the log-fold change and p-value from and creates a row-based plot where each point represents a feature.
}
