% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot-class.R
\docType{methods}
\name{VolcanoPlot-class}
\alias{VolcanoPlot-class}
\alias{VolcanoPlot}
\title{The VolcanoPlot class}
\description{
The VolcanoPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating a volcano plot.
It retrieves the log-fold change and p-value from and creates a row-based plot where each point represents a feature.
}
