% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{de-generics}
\alias{de-generics}
\alias{pValue}
\alias{log2FoldChange}
\alias{averageLog2}
\title{Generics for Differential Expression Results}
\description{
An overview of the generics for accessing common pieces of information in differential expression results.
}
\section{Definitions}{

\itemize{
\item \code{pValue(x)} returns a numeric vector of raw p-values.
\item \code{log2FoldChange(x)} returns a numeric vector of log2-fold-change values.
\item \code{averageLog2(x)} returns a numeric vector of average log2-expression values.
}
}

\examples{
showMethods(pValue)
showMethods(log2FoldChange)
showMethods(averageLog2)
}
\author{
Kevin Rue-Albrecht
}
