% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETable-class.R
\docType{methods}
\name{DETable-class}
\alias{DETable-class}
\alias{DETable}
\alias{initialize,DETable-method}
\alias{.fullName,DETable-method}
\alias{.panelColor,DETable-method}
\title{The DETable class}
\description{
The DETable is a \linkS4class{RowTable} subclass that is dedicated to creating a volcano plot.
It retrieves the log-fold change and p-value from and creates a row-based plot where each point represents a feature.
}
\examples{
x <- DETable()
}
