% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEedgeR-class.R
\docType{methods}
\name{iSEEedgeRResults-class}
\alias{iSEEedgeRResults-class}
\alias{iSEEedgeRResults}
\alias{showAsCell,iSEEedgeRResults-method}
\alias{pValue,iSEEedgeRResults-method}
\alias{log2FoldChange,iSEEedgeRResults-method}
\alias{averageLog2,iSEEedgeRResults-method}
\alias{embedContrastResults,TopTags-method}
\alias{embedContrastResults,iSEEedgeRResults-method}
\title{The iSEEedgeRResults class}
\description{
The \code{iSEEedgeRResults} class is used to provide an common interface to differential expression results produced by the \pkg{edgeR} package.
It provides methods to access common differential expression statistics (e.g., log fold-change, p-value, log2 average abundance).
}
\details{
This class inherits all its slots directly from its parent class \linkS4class{DataFrame}.
}
\section{Constructor}{

\code{iSEEedgeRResults(data, row.names = rownames(data))} creates an instance of a \code{iSEEedgeRResults} class, with:

\describe{
\item{\code{data}}{A \code{data.frame} produced by \code{edgeR::topTags()}.}
\item{\code{row.names}}{The character vector of rownames for the \linkS4class{SummarizedExperiment} object in which the object is to be embedded. Must be a superset of \code{rownames(data)}.}
}
}

\section{Supported methods}{

\itemize{
\item \code{embedContrastResults(x, se, name, ...)} embeds \code{x} in the column \code{name} of \code{rowData(se)[["iSEEde"]]}.
\item \code{pValue(x)} returns the vector of raw p-values.
\item \code{log2FoldChange(x)} returns the vector of log2-fold-change values.
\item \code{averageLog2(x)} returns the vector of average log2-expression values.
}
}

\examples{
library(edgeR)
library(SummarizedExperiment)

##
# From edgeR::topTags() ----
##

# generate raw counts from NB, create list object
y <- matrix(rnbinom(80, size = 1, mu = 10), nrow = 20)
d <- DGEList(counts = y, group = rep(1:2, each = 2), lib.size = rep(c(1000:1001), 2))
rownames(d$counts) <- paste("gene", 1:nrow(d$counts), sep = ".")

# estimate common dispersion and find differences in expression
# here we demonstrate the 'exact' methods, but the use of topTags is
# the same for a GLM analysis
d <- estimateCommonDisp(d)
de <- exactTest(d)

# look at top 10
res <- topTags(de)

##
# iSEEedgeRResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment(assays = list(counts = d$counts))

# Package the results in a iSEEedgeRResults object
iseede_table <- iSEEedgeRResults(res, row.names = rownames(se))

# Store the iSEEedgeRResults object in the SummarizedExperiment rowData
rowData(se)[["iSEEde"]] <- DataFrame(edgeR = I(iseede_table))

se

##
# Methods ----
##

head(pValue(iseede_table))
head(log2FoldChange(iseede_table))
head(averageLog2(iseede_table))
}
\author{
Kevin Rue-Albrecht
}
