% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEElimma-class.R
\docType{methods}
\name{iSEELimmaResults-class}
\alias{iSEELimmaResults-class}
\alias{iSEELimmaResults}
\alias{showAsCell,iSEELimmaResults-method}
\alias{pvalue,iSEELimmaResults-method}
\alias{logfc,iSEELimmaResults-method}
\alias{average,iSEELimmaResults-method}
\title{The iSEELimmaResults class}
\description{
The \code{iSEELimmaResults} class is used to provide an common interface to differential expression results produced by the \pkg{limma} package.
It provides methods to access common differential expression statistics (e.g., log fold-change, p-value, average abundance).
}
\details{
This class inherits all its slots directly from its parent class \linkS4class{DataFrame}.
}
\section{Constructor}{

\code{iSEELimmaResults(data, row.names = rownames(data))} creates an instance of a \code{iSEELimmaResults} class, with:

\describe{
\item{\code{data}}{A \code{data.frame} produced by \code{limma::topTable()}.}
\item{\code{row.names}}{The character vector of rownames for the \linkS4class{SummarizedExperiment} object in which the object is to be embedded. Must be a superset of \code{rownames(data)}.}
}
}

\section{Supported methods}{

\itemize{
\item \code{pvalue(x)} returns the vector of raw p-values.
\item \code{logfc(x)} returns the vector of log2-fold-change values.
\item \code{average(x)} returns the vector of average log2-expression values.
}
}

\examples{
library(limma)
library(SummarizedExperiment)

##
# From limma::lmFit() ----
##

sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- matrix(rnorm(100*6,sd=sd),100,6)
rownames(y) <- paste("Gene",1:100)
y[1:2,4:6] <- y[1:2,4:6] + 2
design <- cbind(Grp1=1, Grp2vs1=c(0,0,0,1,1,1))

fit <- lmFit(y,design)
fit <- eBayes(fit)
tt <- topTable(fit, coef=2)
head(tt)

##
# iSEELimmaResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment(assays = y)

# Package the results in a iSEELimmaResults object
iseede_table <- iSEELimmaResults(tt, row.names=rownames(y))

# Store the iSEELimmaResults object in the SummarizedExperiment rowData
rowData(se)[["iSEEde"]] <- DataFrame(limma=I(iseede_table))

se

##
# Methods ----
##

head(pvalue(iseede_table))
head(logfc(iseede_table))
head(average(iseede_table))
}
\author{
Kevin Rue-Albrecht
}
