% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogFCLogFCPlot-class.R
\docType{methods}
\name{LogFCLogFCPlot-class}
\alias{LogFCLogFCPlot-class}
\alias{LogFCLogFCPlot}
\alias{initialize,LogFCLogFCPlot-method}
\alias{.cacheCommonInfo,LogFCLogFCPlot-method}
\alias{.createObservers,LogFCLogFCPlot-method}
\alias{.defineDataInterface,LogFCLogFCPlot-method}
\alias{.fullName,LogFCLogFCPlot-method}
\alias{.generateDotPlotData,LogFCLogFCPlot-method}
\alias{.panelColor,LogFCLogFCPlot-method}
\alias{.refineParameters,LogFCLogFCPlot-method}
\title{The LogFCLogFCPlot class}
\description{
The LogFCLogFCPlot class is a \linkS4class{RowDataPlot} subclass that is dedicated to comparing the log-fold-change value of two contrasts.
It retrieves the log-fold change of the two selected contrasts and creates a row-based plot where each point represents a feature.
}
\section{Slot overview}{

The following slots control the test procedure:
\itemize{
\item \code{ContrastNameX}, a character scalar indicating the name of the contrast to display on the x-axis.
\item \code{ContrastNameY}, a character scalar indicating the name of the contrast to display on the y-axis.
}

In addition, this class inherits all slots from its parent \linkS4class{RowDotPlot}, \linkS4class{DotPlot}, and \linkS4class{Panel} classes.
}

\examples{
x <- LogFCLogFCPlot()
x
}
