% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot-class.R
\docType{methods}
\name{MAPlot-class}
\alias{MAPlot-class}
\alias{MAPlot}
\alias{initialize,MAPlot-method}
\alias{.cacheCommonInfo,MAPlot-method}
\alias{.createObservers,MAPlot-method}
\alias{.defineDataInterface,MAPlot-method}
\alias{.fullName,MAPlot-method}
\alias{.generateDotPlotData,MAPlot-method}
\alias{.panelColor,MAPlot-method}
\alias{.refineParameters,MAPlot-method}
\title{The MAPlot class}
\description{
The MAPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating an MA plot.
It retrieves the log-fold change (M) and mean average (A) values and creates a row-based plot where each point represents a feature.
}
\section{Slot overview}{

The following slots control the test procedure:
\itemize{
\item \code{ContrastName}, a character scalar indicating the name of the contrast to display.
}

In addition, this class inherits all slots from its parent \linkS4class{RowDotPlot}, \linkS4class{DotPlot}, and \linkS4class{Panel} classes.
}

\examples{
x <- MAPlot()
x
}
