% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot-class.R
\docType{methods}
\name{MAPlot-class}
\alias{MAPlot-class}
\alias{MAPlot}
\alias{initialize,MAPlot-method}
\alias{.cacheCommonInfo,MAPlot-method}
\alias{.createObservers,MAPlot-method}
\alias{.defineDataInterface,MAPlot-method}
\alias{.fullName,MAPlot-method}
\alias{.generateDotPlotData,MAPlot-method}
\alias{.panelColor,MAPlot-method}
\alias{.refineParameters,MAPlot-method}
\title{The MAPlot class}
\description{
The MAPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating a volcano plot.
It retrieves the log-fold change and p-value from and creates a row-based plot where each point represents a feature.
}
\examples{
x <- MAPlot()
x[["ContrastName"]]
x[["ContrastName"]] <- "treated vs control"
}
