% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utils-SummarizedExperiment.R
\docType{methods}
\name{utils-SummarizedExperiment}
\alias{utils-SummarizedExperiment}
\alias{embedResults}
\alias{embedResults,ANY-method}
\alias{embedResultsMethods}
\title{Generics for Embbedding Results into a SummarizedExperiment Object}
\format{
An object of class \code{character} of length 1.
}
\usage{
embedResults(x, se, name, ...)

embedResultsMethods
}
\description{
An overview of the generics for embedding results into a \linkS4class{SummarizedExperiment} object, in a format compatible with \pkg{iSEEde}.
}
\section{Definitions}{

\itemize{
\item \code{embedResults(x, se, name, ...)} embeds the results \code{x} in the \linkS4class{SummarizedExperiment} \code{se}.
}
}

\examples{
embedResultsMethods

showMethods(embedResults)
}
\author{
Kevin Rue-Albrecht
}
\keyword{datasets}
