% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utils-SummarizedExperiment.R
\docType{methods}
\name{utils-SummarizedExperiment}
\alias{utils-SummarizedExperiment}
\alias{embedResults}
\alias{embedResults,ANY-method}
\alias{embedResultsMethods}
\alias{embedResults,data.frame-method}
\title{Generics for Embbedding Results into a SummarizedExperiment Object}
\format{
An object of class \code{character} of length 1.
}
\usage{
embedResults(x, se, name, ...)

embedResultsMethods

\S4method{embedResults}{data.frame}(x, se, name, class, ...)
}
\arguments{
\item{x}{Object to be embedded.}

\item{se}{A \linkS4class{SummarizedExperiment} object.}

\item{name}{Identifier for the embedded object.}

\item{class}{Class to use for embedding \code{x}. Only used when \code{class(x)} does
not uniquely identify the package that generated the object.}
}
\description{
An overview of the generics for embedding results into a \linkS4class{SummarizedExperiment} object, in a format compatible with \pkg{iSEEde}.
}
\section{Definitions}{

\itemize{
\item \code{embedResults(x, se, name, ...)} embeds the results \code{x} in the \linkS4class{SummarizedExperiment} \code{se}.
}
}

\examples{
embedResultsMethods

showMethods(embedResults)
}
\author{
Kevin Rue-Albrecht
}
\keyword{datasets}
