% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot-class.R
\docType{methods}
\name{VolcanoPlot-class}
\alias{VolcanoPlot-class}
\alias{VolcanoPlot}
\alias{initialize,VolcanoPlot-method}
\alias{.cacheCommonInfo,VolcanoPlot-method}
\alias{.createObservers,VolcanoPlot-method}
\alias{.defineDataInterface,VolcanoPlot-method}
\alias{.fullName,VolcanoPlot-method}
\alias{.generateDotPlotData,VolcanoPlot-method}
\alias{.panelColor,VolcanoPlot-method}
\alias{.refineParameters,VolcanoPlot-method}
\title{The VolcanoPlot class}
\description{
The VolcanoPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating a volcano plot.
It retrieves the log-fold change and p-value from and creates a row-based plot where each point represents a feature.
}
\section{Slot overview}{

The following slots control the test procedure:
\itemize{
\item \code{ContrastName}, a character scalar indicating the name of the contrast to display.
}

In addition, this class inherits all slots from its parent \linkS4class{RowDotPlot}, \linkS4class{DotPlot}, and \linkS4class{Panel} classes.
}

\examples{
x <- VolcanoPlot()
x
}
